+++ src/test/kotlin/tornadofx/testapps/InlineFragmentTest.kt
@@ -3,6 +3,7 @@ package tornadofx.testapps
 import javafx.application.Platform
 import javafx.scene.paint.Color
 import javafx.scene.text.FontWeight
+import javafx.stage.StageStyle.DECORATED
 import javafx.stage.StageStyle.UNDECORATED
 import tornadofx.*
 
@@ -18,7 +19,7 @@ class DangerButtonView : View("Do not click the button!") {
                 textFill = Color.RED
             }
             setOnAction {
-                adhocWindow("What do you want?", stageStyle = UNDECORATED) {
+                builderWindow("What do you want?", stageStyle = UNDECORATED, owner = primaryStage) {
                     vbox(10) {
                         style {
                             padding = box(20.px)